Refactor folder placing components
