+++ events-app/src/components/AuthComponents/Login/Login.js
@@ -1,8 +1,8 @@
 import React, {useContext, useState} from 'react';
 import axios from 'axios';
 import {Link, useNavigate} from 'react-router-dom';
-import "../../styles.css"
-import {AuthContext} from "../../AuthContext";
+import "../../../styles.css"
+import {AuthContext} from "../../../AuthContext";
 
 async function login(email, password) {
     try {
