+++ events-app/src/components/AuthComponents/Register/Register.js
@@ -1,8 +1,8 @@
 import React, { useState, useContext } from 'react';
 import axios from 'axios';
 import { useNavigate } from 'react-router-dom';
-import "../../styles.css"
-import { AuthContext } from "../../AuthContext";
+import "../../../styles.css"
+import { AuthContext } from "../../../AuthContext";
 
 async function register(name, email, password) {
     try {
