+++ events-app/src/components/AuthComponents/Register/register.scss
