+++ events-app/src/components/EventComponents/CreateEvent/CreateEvent.js
@@ -1,6 +1,6 @@
 import axios from 'axios';
 import {useState} from 'react';
-import "../../styles.css"
+import "../../../styles.css"
 import {useNavigate} from "react-router-dom";
 
 export default function CreateEvent(props) {
