+++ events-app/src/components/EventComponents/EventCard/event_card.scss
