+++ events-app/src/components/EventComponents/EventDetails/event_details.scss
@@ -1,5 +1,5 @@
 // event_details.scss
-@import "../../variables";
+@import "../../../variables";
 
 .event-details {
   padding: 20px;
