+++ events-app/src/components/EventComponents/Events/Events.js
@@ -1,7 +1,7 @@
 import React, {useState, useEffect} from 'react';
 import axios from 'axios';
 import {Link} from 'react-router-dom';
-import "../../styles.css"
+import "../../../styles.css"
 import EventCard from "../EventCard/EventCard";
 
 function Events() {
