+++ events-app/src/components/EventComponents/Events/events.scss
