+++ events-app/src/components/Layout/Footer/Footer.js
@@ -1,4 +1,4 @@
-import "../../styles.css"
+import "../../../styles.css"
 
 function Footer(){
     return (
