+++ events-app/src/components/Layout/Footer/_footer.scss
