+++ events-app/src/components/Layout/Header/_header.scss
