+++ events-app/src/components/Layout/HomePage/HomePage.js
@@ -1,5 +1,5 @@
 import {Link} from 'react-router-dom';
-import "../../styles.css"
+import "../../../styles.css"
 import { useNavigate } from 'react-router-dom';
