+++ events-app/src/components/Layout/HomePage/_hero_section.scss
@@ -1,5 +1,5 @@
 // _hero_section.scss
-@import "../../variables";
+@import "../../../variables";
 
 .hero-section {
     //background-image: url('../../images/hero-background.jpg');
