+++ events-app/src/components/Layout/HomePage/home.scss
@@ -1,5 +1,5 @@
 // Import common styles
-@import "../../variables";
+@import "../../../variables";
 
 // Import specific styles for hero section
 @import "hero_section";
