+++ events-app/src/styles.scss
@@ -1,12 +1,12 @@
-@import "components/Header/header";
-@import "components/Footer/footer";
-@import "components/EventCard/event_card";
-@import "components/Events/events";
-@import "components/EditEvent/edit_event";
-@import "components/EventDetails/event_details";
-@import "components/CreateEvent/create_event";
+@import "components/Layout/Header/header";
+@import "components/Layout/Footer/footer";
+@import "components/EventComponents/EventCard/event_card";
+@import "components/EventComponents/Events/events";
+@import "components/EventComponents/EditEvent/edit_event";
+@import "components/EventComponents/EventDetails/event_details";
+@import "components/EventComponents/CreateEvent/create_event";
 @import "variables";
-@import "components/Login/login";
-@import "components/Register/register";
-@import "components/HomePage/hero_section";
-@import "components/HomePage/home";
+@import "components/AuthComponents/Login/login";
+@import "components/AuthComponents/Register/register";
+@import "components/Layout/HomePage/hero_section";
+@import "components/Layout/HomePage/home";
